[#5]fix: bug fix caused by naming mistake between 'thumnails' and 'thumbnails'
